removed store interface and renamed ipfs lite protocol impl
